+++ CHANGES.md
@@ -11,6 +11,8 @@
   (#2686)
 - No longer color diff headers white as it's unreadable in light themed terminals
   (#2691)
+- No longer attempt to write grammar table cache on import emitting confusing log
+  messages (#2701)
 
 ## 21.12b0